Changing hashing interface to use size_t

+++ BloomFilter/src/Hash.h
@@ -6,5 +6,5 @@
 class Hash {
 public:
 	std::array<uint32_t, 2> hash(const std::string& input) const;
-	std::vector<uint32_t> hash(const std::string& input, unsigned int iterations, unsigned int max) const;
+	std::vector<size_t> hash(const std::string& input, unsigned int iterations, size_t max) const;
 };
